+++ implementation/rocksdb/db/db_impl_switcher/db_impl_switcher.cc
@@ -1,6 +1,6 @@
 #include "db/db_impl/db_impl.h"
 #ifdef TROPODB_PLUGIN_ENABLED
-#include "db/zns_impl/db_impl_zns.h"
+#include "db/zns_impl/tropodb_impl.h"
 #endif
 #include "rocksdb/db.h"
 #include "rocksdb/file_checksum.h"
