+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/zns_impl/db_impl_zns.h"
+#include "db/zns_impl/tropodb_impl.h"
 
 #include <algorithm>
 #include <atomic>
@@ -18,7 +18,7 @@
 #include "db/column_family.h"
 #include "db/memtable.h"
 #include "db/write_batch_internal.h"
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/index/zns_compaction.h"
 #include "db/zns_impl/index/zns_version.h"
 #include "db/zns_impl/index/zns_version_set.h"
