+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl_background.cc
@@ -15,8 +15,8 @@
 #include "db/column_family.h"
 #include "db/memtable.h"
 #include "db/write_batch_internal.h"
-#include "db/zns_impl/config.h"
-#include "db/zns_impl/db_impl_zns.h"
+#include "db/zns_impl/tropodb_config.h"
+#include "db/zns_impl/tropodb_impl.h"
 #include "db/zns_impl/index/zns_compaction.h"
 #include "db/zns_impl/index/zns_version.h"
 #include "db/zns_impl/index/zns_version_set.h"
