+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl_diagnostics.cc
@@ -11,8 +11,8 @@
 #include <string>
 #include <vector>
 
-#include "db/zns_impl/config.h"
-#include "db/zns_impl/db_impl_zns.h"
+#include "db/zns_impl/tropodb_config.h"
+#include "db/zns_impl/tropodb_impl.h"
 #include "db/zns_impl/utils/tropodb_diagnostics.h"
 
 namespace ROCKSDB_NAMESPACE {
