+++ implementation/rocksdb/db/zns_impl/persistence/zns_committer.cc
@@ -1,7 +1,7 @@
 #include "db/zns_impl/persistence/zns_committer.h"
 
 #include "db/write_batch_internal.h"
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 #include "rocksdb/slice.h"
