+++ implementation/rocksdb/db/zns_impl/persistence/zns_wal_manager.ipp
@@ -7,7 +7,7 @@
 #include <iostream>
 
 #include "db/write_batch_internal.h"
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/memtable/zns_memtable.h"
 #include "db/zns_impl/persistence/zns_committer.h"
