+++ implementation/rocksdb/db/zns_impl/table/ln_zns_sstable.h
@@ -3,7 +3,7 @@
 #ifndef LN_ZNS_SSTABLE_H
 #define LN_ZNS_SSTABLE_H
 
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/memtable/zns_memtable.h"
 #include "db/zns_impl/table/zns_sstable.h"
 #include "db/zns_impl/table/zns_sstable_builder.h"
