+++ implementation/rocksdb/db/zns_impl/table/zns_sstable_builder.cc
@@ -1,6 +1,6 @@
 #include "db/zns_impl/table/zns_sstable_builder.h"
 
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/table/ln_zns_sstable.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
