+++ implementation/rocksdb/db/zns_impl/table/zns_sstable_manager.cc
@@ -3,7 +3,7 @@
 #include <iomanip>
 #include <iostream>
 
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/memtable/zns_memtable.h"
 #include "db/zns_impl/table/iterators/sstable_ln_iterator.h"
