+++ implementation/rocksdb/db/zns_impl/table/zns_sstable_manager.h
@@ -5,7 +5,7 @@
 
 #include <optional>
 
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/utils/tropodb_diagnostics.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/memtable/zns_memtable.h"
