+++ implementation/rocksdb/db/zns_impl/tropodb_impl.h
@@ -24,7 +24,7 @@
 #include <utility>
 #include <vector>
 
-#include "db/zns_impl/config.h"
+#include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/index/zns_version.h"
 #include "db/zns_impl/index/zns_version_set.h"
 #include "db/zns_impl/io/szd_port.h"
